requirements.txt
django==6.0
python-decouple
djangorestframework
django-cors-headers
djangorestframework-simplejwt